Refactor test learner into separate file
